snapshots: testsuite: unmount snapshot before committing
Otherwise there is no guarantee that the written data is
flushed back to disk when committing. For a block device
based snapshotter, it can cause data inconsistency if the
committed snapshot is quickly used to create another snapshot
before kernel flushed the dirty blocks.

To workaround it, unmount the snapshot before committing it.
unmount can make sure page caches are all flushed to disk.

Signed-off-by: Peng Tao <bergwolf@hyper.sh>